Fix unity 5.6 compilation issue
The behaviour editor cannot reference this script in unity56 but we can merge the two files and use compiler directives to achieve the same effect
