--- unity/PackageProject/Assets/Standard Assets/Editor/BugsnagBehaviourEditor.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 335b73f255bf2425395884747155c198
-timeCreated: 1532121183
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
